requirements.txt
openai>=1.40
python-dotenv>=1.0
pydantic>=2.7
PyYAML>=6.0
langchain-openai>=0.1.25
langchain-core>=0.2.30
langchain-community>=0.2.10
